# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Self Deepsearch Platform API
  version: 0.1.0
  description: Release A contract baseline. Automatic collection is not enabled.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /admin/v1/media/manifests:
    post:
      operationId: adminRegisterMediaManifest
      description: Registers already-uploaded S3 renditions only after Beijing backup and rights review. The immutable manifest, entity link, audit and durable cache refresh event commit in one transaction; delivery of cache refresh is asynchronous. Draft preparation is allowed, but takedown or merged targets are rejected. The server does not fetch source or public URLs. Requires admin or owner. This operation adds a new asset; it does not replace an existing primary image.
      security: [{ sessionCookie: [], recentAuthCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MediaManifestRequest" }
      responses:
        "201":
          description: Immutable media manifest registered and linked, with a durable cache refresh event queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaManifestResponse" }
        "409": { description: "Target is withdrawn or merged, or storage key, URL, rendition or primary image conflicts with an existing asset" }
        "503": { description: Registration could not be confirmed. No success is returned when the transaction or cache-event enqueue fails; inspect the asset before retrying after an uncertain commit outcome. }
...
  /admin/v1/media/upload-control:
    get:
      operationId: adminGetMediaUploadControl
      description: Requires admin or owner. Returns nullable authenticated Beijing upload snapshot and latest ten durable command outcomes. A fresh snapshot is not permission to resume; pending/running/uncertain does not mean applied. No endpoint or signing secret is exposed.
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Current execution snapshot and command history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUploadOverview" }
        "401": { description: Login required }
        "403": { description: Role is not allowed }
        "503": { description: Snapshot unavailable }
  /admin/v1/media/upload-control/commands:
    post:
      operationId: adminSubmitMediaUploadCommand
      description: Requires admin or owner and recent password confirmation. Queues one explicit pause or resume against exact target, epoch, generation and observed timestamp. Confirmation is mandatory. Requests expire within five minutes. Same actor and idempotency key with identical content returns the previous outcome; conflicting or concurrently open requests are rejected. Worker checks operator authority and resume safety before dispatch. Network uncertainty remains visible until authenticated reconciliation; the API never claims remote application at enqueue time.
      security: [{ sessionCookie: [], recentAuthCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MediaUploadInput" }
      responses:
        "202":
          description: Durable pending command, not yet executed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUploadResult" }
        "200":
          description: Identical idempotent replay of an existing command
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUploadResult" }
        "400": { description: Invalid snapshot, mode, identity, reason or confirmation }
        "401": { description: Login or recent password confirmation required }
        "403": { description: Operator role is not allowed or no longer active }
        "404": { description: Upload executor has not initialized its durable snapshot }
        "409": { description: Snapshot changed, unavailable or stale; executor busy; another command open; or changed idempotency payload }
        "503": { description: Enqueue commit not confirmed; preserve the key and inspect or retry exactly }
  /admin/v1/media/primary:
    get:
      operationId: adminGetPrimaryMedia
      description: Reads the current primary association and parent status in one snapshot, including draft and hidden parents. Requires admin or owner. Does not expose private storage locations.
      security: [{ sessionCookie: [] }]
      parameters:
        - in: query
          name: entity_type
          required: true
          schema: { enum: [work, performer] }
        - in: query
          name: entity_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Current primary, or null if none
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PrimaryMediaStateResponse" }
        "400": { description: Invalid entity identity }
        "401": { description: Login required }
        "403": { description: Role is not allowed }
        "404": { description: Parent does not exist }
        "503": { description: Read could not be completed }
  /admin/v1/media/primary/replace:
    post:
      operationId: adminReplacePrimaryMedia
      description: Requires admin or owner and recent password confirmation. Compares expected_asset_id under the parent lock and registers a different immutable primary manifest. New media, old link withdrawal, audit, cache refresh and retirement tasks commit in one transaction. Shared old assets remain available. Unshared old assets are hidden and public deletions are due immediately; private masters are retained for 30 days before deletion is due. Rights takedown overrides retention. Physical deletion and cache delivery are asynchronous, not confirmed by this response. No external URL is fetched.
      security: [{ sessionCookie: [], recentAuthCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReplacePrimaryMediaRequest" }
      responses:
        "201":
          description: Primary switched; remote deletion is not yet confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReplacePrimaryMediaResponse" }
        "400": { description: "Invalid manifest, missing expected asset, same asset, or not a primary image" }
        "401": { description: Login or recent password confirmation required }
        "403": { description: Role is not allowed }
        "404": { description: Parent does not exist }
        "409": { description: "Primary changed or disappeared, parent or old asset is withdrawn, or immutable new manifest conflicts. Reload the primary and explicitly confirm again; never automatically overwrite." }
        "503": { description: Transaction could not be confirmed. Preserve the manifest and inspect the current primary before retrying an uncertain commit. }
...
components:
  securitySchemes:
    metricsBearer:
      type: http
      scheme: bearer
    sessionCookie:
      type: apiKey
      in: cookie
      name: sd_session
    recentAuthCookie:
      type: apiKey
      in: cookie
      name: sd_reauth
...
  schemas:
    MediaUploadReceipt:
      type: object
      additionalProperties: false
      required: [status, epoch, generation, mode, command_id, applied_at]
      properties:
        status: { type: string, enum: [observed, applied, replayed] }
        epoch: { type: string, description: Empty only for an uninitialized paused executor }
        generation: { type: integer, minimum: 0, maximum: 9007199254740991 }
        mode: { type: string, enum: [paused, enabled] }
        command_id: { type: string, description: Empty only for generation zero }
        applied_at: { type: string, description: UTC whole seconds; empty only for generation zero }
    MediaUploadState:
      type: object
      required: [target_ref, receipt, last_success_at, updated_at, error_code, fresh]
      properties:
        target_ref: { type: string, pattern: "^[a-f0-9]{64}$" }
        receipt: { anyOf: [{ $ref: "#/components/schemas/MediaUploadReceipt" }, { type: "null" }] }
        last_success_at: { type: [string, "null"], format: date-time }
        updated_at: { type: string, format: date-time }
        error_code: { type: [string, "null"] }
        fresh: { type: boolean }
    MediaUploadCommand:
      type: object
      required: [command_id, actor_id, mode, reason, status, created_at, expires_at, first_dispatched_at, completed_at, dispatch_count, error_code, receipt, idempotent_replay]
      properties:
        command_id: { type: string, format: uuid }
        actor_id: { type: string, format: uuid }
        mode: { type: string, enum: [paused, enabled] }
        reason: { type: string }
        status: { type: string, enum: [pending, running, uncertain, applied, rejected, superseded] }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        first_dispatched_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
        dispatch_count: { type: integer, minimum: 0 }
        error_code: { type: [string, "null"] }
        receipt: { anyOf: [{ $ref: "#/components/schemas/MediaUploadReceipt" }, { type: "null" }] }
        idempotent_replay: { type: boolean }
    MediaUploadOverview:
      type: object
      required: [state, commands]
      properties:
        state: { anyOf: [{ $ref: "#/components/schemas/MediaUploadState" }, { type: "null" }] }
        commands: { type: array, maxItems: 10, items: { $ref: "#/components/schemas/MediaUploadCommand" } }
    MediaUploadInput:
      type: object
      additionalProperties: false
      required: [idempotency_key, target_ref, expected_epoch, expected_generation, expected_observed_at, mode, reason, confirmed]
      properties:
        idempotency_key: { type: string, format: uuid, description: Canonical lowercase UUID v4 }
        target_ref: { type: string, pattern: "^[a-f0-9]{64}$" }
        expected_epoch: { type: string }
        expected_generation: { type: integer, minimum: 0, maximum: 9007199254740990 }
        expected_observed_at: { type: string, format: date-time }
        mode: { type: string, enum: [paused, enabled] }
        reason: { type: string, minLength: 2, maxLength: 1000 }
        confirmed: { type: boolean, enum: [true] }
    MediaUploadResult:
      type: object
      required: [command]
      properties:
        command: { $ref: "#/components/schemas/MediaUploadCommand" }
...
    MediaManifestObject:
      type: object
      additionalProperties: false
      required: [rendition, storage_scope, storage_key, backup_path, public_url, sha256, mime_type, width, height, byte_size]
      properties:
        rendition: { enum: [master, w320, w640, w960] }
        storage_scope: { enum: [private, public] }
        storage_key: { type: string, minLength: 1, maxLength: 1024 }
        backup_path: { type: string, minLength: 1, maxLength: 2048 }
        public_url: { type: [string, "null"], format: uri, pattern: "^https://" }
        sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
        mime_type: { const: image/webp }
        width: { type: integer, minimum: 1, maximum: 20000 }
        height: { type: integer, minimum: 1, maximum: 20000 }
        byte_size: { type: integer, minimum: 1, maximum: 10485760 }
    MediaManifestRequest:
      type: object
      additionalProperties: false
      required: [asset_id, entity_type, entity_id, asset_type, source_type, checked_at, confidence, purpose, position, is_primary, reason, tool_version, objects]
      properties:
        asset_id: { type: string, format: uuid }
        entity_type: { enum: [work, performer] }
        entity_id: { type: string, format: uuid }
        asset_type: { enum: [work_image, performer_avatar] }
        source_type: { type: string, minLength: 1, maxLength: 100 }
        source_url: { type: [string, "null"], format: uri }
        checked_at: { type: string, format: date-time }
        confidence: { type: number, minimum: 0, maximum: 1 }
        purpose: { enum: [cover, gallery, avatar], description: "Work cover and performer avatar are the primary slot; work gallery is a secondary slot." }
        position: { type: integer, minimum: 0, maximum: 3, description: "Primary cover/avatar must use 0; work gallery must use 1, 2 or 3." }
        is_primary: { type: boolean, description: "Must be true for cover/avatar and false for gallery." }
        reason: { type: string, minLength: 2, maxLength: 1000 }
        tool_version: { const: media-python/1 }
        objects:
          type: array
          minItems: 4
          maxItems: 4
          items: { $ref: "#/components/schemas/MediaManifestObject" }
    MediaManifestResponse:
      type: object
      required: [asset_id, entity_media_id, version, object_count, status, created_at]
      properties:
        asset_id: { type: string, format: uuid }
        entity_media_id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        object_count: { type: integer, minimum: 4, maximum: 4 }
        status: { const: published }
        created_at: { type: string, format: date-time }
    PrimaryMedia:
      type: object
      required: [asset_id, entity_media_id, purpose]
      properties:
        asset_id: { type: string, format: uuid }
        entity_media_id: { type: string, format: uuid }
        purpose: { enum: [cover, gallery, avatar] }
    PrimaryMediaStateResponse:
      type: object
      required: [entity_type, entity_id, entity_status, primary]
      properties:
        entity_type: { enum: [work, performer] }
        entity_id: { type: string, format: uuid }
        entity_status: { enum: [draft, reviewing, approved, published, hidden, takedown, merged] }
        primary:
          anyOf:
            - { $ref: "#/components/schemas/PrimaryMedia" }
            - { type: "null" }
    ReplacePrimaryMediaRequest:
      type: object
      additionalProperties: false
      required: [expected_asset_id, manifest]
      properties:
        expected_asset_id: { type: string, format: uuid }
        manifest: { $ref: "#/components/schemas/MediaManifestRequest" }
    ReplacePrimaryMediaResponse:
      type: object
      required: [media, previous_asset_id, old_asset_retired, private_retained_until]
      properties:
        media: { $ref: "#/components/schemas/MediaManifestResponse" }
        previous_asset_id: { type: string, format: uuid }
        old_asset_retired: { type: boolean }
        private_retained_until: { type: [string, "null"], format: date-time }
...